Attempt at 2024 Day 6: Guard Gallivant
I completed Part 1 fine, but Part 2 has me beat. My messy solution is
slow but sound as near as I can tell. It works on the example input.
But it fails on the actual input. There are some 1800 iterations which
may be buggy, so I am at a loss for how to move forward on this one.
Maybe I will revisit it later in the month.