Solve 2024 day 11: Plutonian Pebbles
A very rough part two. I spent a long time trying to figure out an
approach to tame the brute force array overruns. There is definitely
something to the way the numbers repeat in semi-predictable ways and
how in the end all you need is the count not the actual numbers...

...or I could just split the task in half and avoid the array overruns.
A very silly solution that took awhile to run, making mistakes costly.
Oh well.